move power_grid_model.hpp
