+++ power_grid_model_c/power_grid_model/include/power_grid_model/auxiliary/meta_gen/update.hpp
@@ -12,7 +12,7 @@
 #include "gen_getters.hpp" // NOLINT
 
 #include "../../enum.hpp"               // NOLINT
-#include "../../power_grid_model.hpp"   // NOLINT
+#include "../../common/power_grid_model.hpp"   // NOLINT
 #include "../../three_phase_tensor.hpp" // NOLINT
 #include "../meta_data.hpp"             // NOLINT
 #include "../update.hpp" // NOLINT
