+++ power_grid_model_c/power_grid_model/include/power_grid_model/auxiliary/serialization/serializer.hpp
@@ -6,8 +6,8 @@
 #ifndef POWER_GRID_MODEL_AUXILIARY_SERIALIZTION_SERIALIZER_HPP
 #define POWER_GRID_MODEL_AUXILIARY_SERIALIZTION_SERIALIZER_HPP
 
+#include "../../common/power_grid_model.hpp"
 #include "../../exception.hpp"
-#include "../../power_grid_model.hpp"
 #include "../dataset_handler.hpp"
 #include "../meta_data.hpp"
 #include "../meta_data_gen.hpp"
