+++ power_grid_model_c/power_grid_model/include/power_grid_model/calculation_parameters.hpp
@@ -6,9 +6,9 @@
 #ifndef POWER_GRID_MODEL_CALCULATION_PARAMETERS_HPP
 #define POWER_GRID_MODEL_CALCULATION_PARAMETERS_HPP
 
+#include "common/power_grid_model.hpp"
 #include "enum.hpp"
 #include "grouped_index_vector.hpp"
-#include "power_grid_model.hpp"
 #include "three_phase_tensor.hpp"
 
 namespace power_grid_model {
