+++ power_grid_model_c/power_grid_model/include/power_grid_model/component/node.hpp
@@ -12,7 +12,7 @@
 #include "../auxiliary/output.hpp"
 #include "../auxiliary/update.hpp"
 #include "../calculation_parameters.hpp"
-#include "../power_grid_model.hpp"
+#include "../common/power_grid_model.hpp"
 #include "../three_phase_tensor.hpp"
 
 namespace power_grid_model {
