+++ power_grid_model_c/power_grid_model/include/power_grid_model/component/power_sensor.hpp
@@ -9,9 +9,9 @@
 #include "sensor.hpp"
 
 #include "../calculation_parameters.hpp"
+#include "../common/power_grid_model.hpp"
 #include "../enum.hpp"
 #include "../exception.hpp"
-#include "../power_grid_model.hpp"
 
 namespace power_grid_model {
