+++ power_grid_model_c/power_grid_model/include/power_grid_model/component/transformer.hpp
@@ -13,8 +13,8 @@
 #include "../auxiliary/output.hpp"
 #include "../auxiliary/update.hpp"
 #include "../calculation_parameters.hpp"
+#include "../common/power_grid_model.hpp"
 #include "../exception.hpp"
-#include "../power_grid_model.hpp"
 #include "../three_phase_tensor.hpp"
 
 namespace power_grid_model {
