+++ power_grid_model_c/power_grid_model/include/power_grid_model/component/voltage_sensor.hpp
@@ -9,8 +9,8 @@
 #include "sensor.hpp"
 
 #include "../calculation_parameters.hpp"
+#include "../common/power_grid_model.hpp"
 #include "../exception.hpp"
-#include "../power_grid_model.hpp"
 #include "../three_phase_tensor.hpp"
 
 #include <limits>
