+++ power_grid_model_c/power_grid_model/include/power_grid_model/container.hpp
@@ -8,8 +8,8 @@
 
 // container for multiple components
 
+#include "common/power_grid_model.hpp"
 #include "exception.hpp"
-#include "power_grid_model.hpp"
 
 #include <boost/iterator/iterator_facade.hpp>
 #include <boost/range.hpp>
