+++ power_grid_model_c/power_grid_model/include/power_grid_model/grouped_index_vector.hpp
@@ -6,7 +6,7 @@
 #ifndef POWER_GRID_MODEL_SPARSE_IDX_VECTOR_HPP
 #define POWER_GRID_MODEL_SPARSE_IDX_VECTOR_HPP
 
-#include "power_grid_model.hpp"
+#include "common/power_grid_model.hpp"
 #include "typing.hpp"
 
 #include <boost/iterator/iterator_facade.hpp>
