+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/linear_pf_solver.hpp
@@ -35,8 +35,8 @@ if there are sources
 #include "y_bus.hpp"
 
 #include "../calculation_parameters.hpp"
+#include "../common/power_grid_model.hpp"
 #include "../exception.hpp"
-#include "../power_grid_model.hpp"
 #include "../three_phase_tensor.hpp"
 #include "../timer.hpp"
