+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/sparse_lu_solver.hpp
@@ -6,8 +6,8 @@
 #ifndef POWER_GRID_MODEL_MATH_SOLVER_SPARSE_LU_SOLVER_HPP
 #define POWER_GRID_MODEL_MATH_SOLVER_SPARSE_LU_SOLVER_HPP
 
+#include "../common/power_grid_model.hpp"
 #include "../exception.hpp"
-#include "../power_grid_model.hpp"
 #include "../three_phase_tensor.hpp"
 #include "../typing.hpp"
