+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/y_bus.hpp
@@ -7,7 +7,7 @@
 #define POWER_GRID_MODEL_MATH_SOLVER_Y_BUS_HPP
 
 #include "../calculation_parameters.hpp"
-#include "../power_grid_model.hpp"
+#include "../common/power_grid_model.hpp"
 #include "../three_phase_tensor.hpp"
 
 #include <algorithm>
