+++ power_grid_model_c/power_grid_model/include/power_grid_model/three_phase_tensor.hpp
@@ -6,7 +6,7 @@
 #ifndef POWER_GRID_MODEL_THREE_PHASE_TENSOR_HPP
 #define POWER_GRID_MODEL_THREE_PHASE_TENSOR_HPP
 
-#include "power_grid_model.hpp"
+#include "common/power_grid_model.hpp"
 
 // eigen properties
 #include <Eigen/Dense>
